Add DataManager
To reduce complexity of the Game class, move functionality related to
participant data and schemas to a DataManager class. This change also
required changes to the schema-util, which now has to check for
automatic schema in the DataManager class as well as the Game class
(for backwards compatibility). Finally, the code for intercepting WebGl
calls was moved from Game class to the WebGlInfo class.