Update dependencies and docker config
The prior docker compose config stopped working, perhaps due to changes
in most recent docker. Now, the integration-test script in the root
package.json uses "docker compose" rather than "docker-compose", and the
images for the services are now built in two separate steps: the first
builds the m2c2kit image because it is needed by services in the second
step. Also, perhaps due to changes in docker or node, some services
are exiting with code 137. This is probably because they are not
responding to SIGTERM and thus 137 means the services were were
terminated by SIGKILL. This is OK because we depend only on the exit
code from the m2c2kit service, which runs the integration tests.